// src/interfaces/Books.ts
export interface IBook {
  id: number
  title: string
  authors: string
  average_rating: number
  ratings_count: number
  image_url: string
  isbn: string
  language_code: string
  original_publication_year: number
  original_title: string
  ratings_1: number
  ratings_2: number
  ratings_3: number
  ratings_4: number
  ratings_5: number
}